Tema08 - Seguridad con JWT/Clase24_28-11-2025/api-gastos-flask/requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-jwt-extended==4.7.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
cryptography
flask-mail